requirements.txt
Package            Version
------------------ ---------
certifi            2022.9.24
charset-normalizer 2.1.1
contourpy          1.0.6
cycler             0.11.0
fonttools          4.38.0
idna               3.4
kiwisolver         1.4.4
matplotlib         3.6.2
numpy              1.23.4
opencv-python      4.6.0.66
packaging          21.3
Pillow             9.3.0
pip                22.2.2
pyparsing          3.0.9
python-dateutil    2.8.2
requests           2.28.1
scipy              1.9.3
setuptools         63.2.0
six                1.16.0
torch              1.11.0
torchvision        0.12.0
tqdm               4.64.1
typing_extensions  4.4.0
urllib3            1.26.12
